backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
anthropic
python-multipart
python-dotenv
alembic
numpy
openai
scikit-learn
chromadb
sentence-transformers
spacy
presidio_analyzer
presidio_anonymizer
pdfplumber
python-docx
pymupdf